refactor: [GSW-1683] services/converter architect
