+++ packages/web/src/layouts/pool/pool-remove/containers/remove-liquidity-container/RemoveLiquidityContainer.tsx
@@ -5,7 +5,7 @@ import useCustomRouter from "@hooks/common/use-custom-router";
 import { usePositionData } from "@hooks/pool/data/use-position-data";
 import { useWallet } from "@hooks/wallet/data/use-wallet";
 import { PoolPositionModel } from "@models/position/pool-position-model";
-import { PositionConverter } from "@services/position";
+import { PositionConverter } from "@services/converters/position";
 
 import RemoveLiquidity from "../../components/remove-liquidity/RemoveLiquidity";
 import { useRemovePositionModal } from "@hooks/pool/ui/use-remove-position-modal";
