+++ packages/web/src/services/converters/position/position.converter.ts
@@ -1,4 +1,4 @@
-import { AmountConverter } from "@services/converters";
+import { AmountConverter } from "@services/converters/common/amount";
 
 import { PoolPositionModel } from "@models/position/pool-position-model";
